+++ app/src/test/java/es/npatarino/android/gotchallenge/houses/data/source/local/mapper/BddHouseMapperTest.java
@@ -1,4 +1,4 @@
-package es.npatarino.android.gotchallenge.data.source.local.entities.mapper;
+package es.npatarino.android.gotchallenge.houses.data.source.local.mapper;
 
 import android.support.annotation.NonNull;
 
@@ -9,7 +9,6 @@ import java.util.List;
 
 import es.npatarino.android.gotchallenge.BuildConfig;
 import es.npatarino.android.gotchallenge.houses.data.source.local.entities.BddHouse;
-import es.npatarino.android.gotchallenge.houses.data.source.local.mapper.BddHouseMapper;
 import es.npatarino.android.gotchallenge.houses.domain.model.House;
 
 import static org.hamcrest.core.Is.is;
